py-flask-sqlalchemy/requirements.txt
asposestorage==1.0.2
Flask==2.3.2
flask_restful==0.3.10
flask_sqlalchemy==3.0.5
